feat: rota sale

+++ back-end/src/database/migrations/03-create-products.js
@@ -15,7 +15,7 @@ module.exports = {
         },
         price: {
           allowNull: false,
-          type: Sequelize.DOUBLE,
+          type: Sequelize.DECIMAL(9, 2),
         },
         urlImage: {
           allowNull: false,
